requirements.txt
flask
flask-cors
python-dotenv
supabase
requests